requirements.txt
requests~=2.24.0
peewee~=3.13.3
paho-mqtt~=1.5.1
toml~=0.10.1